@@ -8,7 +8,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import com.stfalcon.sample.R
 import com.stfalcon.sample.common.extensions.sendShareIntent
 import com.stfalcon.sample.common.models.Poster
-import kotlinx.android.synthetic.main.view_poster_overlay.view.*
+import com.stfalcon.sample.databinding.ViewPosterOverlayBinding
 
 class PosterOverlayView @JvmOverloads constructor(
     context: Context,
@@ -16,16 +16,20 @@ class PosterOverlayView @JvmOverloads constructor(
     defStyleAttr: Int = 0
 ) : ConstraintLayout(context, attrs, defStyleAttr) {
 
+    private val binding: ViewPosterOverlayBinding
+
     var onDeleteClick: (Poster) -> Unit = {}
 
     init {
         View.inflate(context, R.layout.view_poster_overlay, this)
+        binding = ViewPosterOverlayBinding.bind(this)
+
         setBackgroundColor(Color.TRANSPARENT)
     }
 
     fun update(poster: Poster) {
-        posterOverlayDescriptionText.text = poster.description
-        posterOverlayShareButton.setOnClickListener { context.sendShareIntent(poster.url) }
-        posterOverlayDeleteButton.setOnClickListener { onDeleteClick(poster) }
+        binding.posterOverlayDescriptionText.text = poster.description
+        binding.posterOverlayShareButton.setOnClickListener { context.sendShareIntent(poster.url) }
+        binding.posterOverlayDeleteButton.setOnClickListener { onDeleteClick(poster) }
     }
 }
