@@ -8,7 +8,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import com.stfalcon.sample.R
 import com.stfalcon.sample.common.models.Demo
 import com.stfalcon.sample.common.models.Poster
-import kotlinx.android.synthetic.main.view_posters_grid.view.*
+import com.stfalcon.sample.databinding.ViewPostersGridBinding
 
 class PostersGridView @JvmOverloads constructor(
     context: Context,
@@ -16,24 +16,26 @@ class PostersGridView @JvmOverloads constructor(
     defStyleAttr: Int = 0
 ) : ConstraintLayout(context, attrs, defStyleAttr) {
 
+    private lateinit var binding : ViewPostersGridBinding
     var imageLoader: ((ImageView, Poster?) -> Unit)? = null
     var onPosterClick: ((Int, ImageView) -> Unit)? = null
 
     val imageViews by lazy {
         mapOf<Int, ImageView>(
-            0 to postersFirstImage,
-            1 to postersSecondImage,
-            2 to postersThirdImage,
-            3 to postersFourthImage,
-            4 to postersFifthImage,
-            5 to postersSixthImage,
-            6 to postersSeventhImage,
-            7 to postersEighthImage,
-            8 to postersNinthImage)
+            0 to binding.postersFirstImage,
+            1 to binding.postersSecondImage,
+            2 to binding.postersThirdImage,
+            3 to binding.postersFourthImage,
+            4 to binding.postersFifthImage,
+            5 to binding.postersSixthImage,
+            6 to binding.postersSeventhImage,
+            7 to binding.postersEighthImage,
+            8 to binding.postersNinthImage)
     }
 
     init {
         View.inflate(context, R.layout.view_posters_grid, this)
+        binding = ViewPostersGridBinding.bind(this)
     }
 
     override fun onAttachedToWindow() {
