@@ -7,9 +7,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.stfalcon.sample.R
-import kotlinx.android.synthetic.main.fragment_demo_card.*
+import com.stfalcon.sample.databinding.FragmentDemoCardBinding
 
 class DemoCardFragment : Fragment() {
+
+    private lateinit var binding: FragmentDemoCardBinding
 
     companion object {
         private const val ARG_ACTION_ID = "ARG_ACTION_ID"
@@ -41,14 +43,17 @@ class DemoCardFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = inflater.inflate(R.layout.fragment_demo_card, container, false)
+    ): View {
+        binding = FragmentDemoCardBinding.inflate(inflater)
+        return binding.root
+    }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        demoCardTitleText.text = title
-        demoCardDescriptionText.text = description
-        demoCardActionButton.setOnClickListener { cardActionListener?.onCardAction(actionId) }
+        binding.demoCardTitleText.text = title
+        binding.demoCardDescriptionText.text = description
+        binding.demoCardActionButton.setOnClickListener { cardActionListener?.onCardAction(actionId) }
     }
 
     override fun onAttach(context: Context) {
