@@ -26,11 +26,11 @@ internal fun ViewPropertyAnimator.setAnimatorListener(
 ) = this.setListener(
     object : AnimatorListenerAdapter() {
 
-        override fun onAnimationEnd(animation: Animator?) {
+        override fun onAnimationEnd(animation: Animator) {
             onAnimationEnd?.invoke(animation)
         }
 
-        override fun onAnimationStart(animation: Animator?) {
+        override fun onAnimationStart(animation: Animator) {
             onAnimationStart?.invoke(animation)
         }
     })
